Refactored Day 13 to new format
