--- 2019/Day13/Solution/Day13.csproj
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
-    <UserSecretsId>af2ad03d-386b-4b2e-b3b5-ccb5f138b698</UserSecretsId>
-    <Nullable>enable</Nullable>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="MoreLinq" Version="3.3.2" />
-    <PackageReference Include="NAoCHelper" Version="1.2.5" />
-  </ItemGroup>
-</Project>
