--- 2019/Day13/Solution/appsettings.json
@@ -1,5 +1,0 @@
-{
-    "Secrets": {
-        "Cookie": "Configured in User Secrets"
-    }
-}
