--- 2019/Day13/Tests/ArcadeCabinetTests.cs
@@ -1,20 +1,0 @@
-using System.Numerics;
-using Day13;
-using NUnit.Framework;
-
-namespace Day13Tests
-{
-    public class ArcadeCabinetTests
-    {
-        [TestCase]
-        public void CanGenerateGridSuccessfully()
-        {
-            // Arrange
-
-            // Act
-
-            // Assert
-            Assert.Pass();
-        }
-    }
-}
