--- 2019/Day13/Tests/OperationTests.cs
@@ -1,31 +1,0 @@
-using System.Collections.Generic;
-using System.Numerics;
-using Day13;
-using NUnit.Framework;
-using Shouldly;
-
-namespace Day13Tests
-{
-    public class OperationTests
-    {
-        [TestCaseSource(nameof(ConstructorCases))]
-        public void CanSuccessfullyConstructOperationFromInteger(BigInteger input, OpCode expectedOpCode, Mode expectedMode1, Mode expectedMode2, Mode expectedMode3)
-        {
-            // Arrange
-
-            // Act
-            var operation = new Operation(input);
-
-            // Assert
-            operation.OpCode.ShouldBe(expectedOpCode);
-            operation.FirstParameterMode.ShouldBe(expectedMode1);
-            operation.SecondParameterMode.ShouldBe(expectedMode2);
-            operation.ThirdParameterMode.ShouldBe(expectedMode3);
-        }
-
-        private static IEnumerable<object[]> ConstructorCases()
-        {
-            yield return new object[] { new BigInteger(1002), OpCode.Multiply, Mode.Position, Mode.Immediate, Mode.Position };
-        }
-    }
-}
