+++ 2019/Solutions/Day13/Enums.cs
@@ -1,0 +1,9 @@
+﻿namespace Solutions.Day13
+{
+	public enum Direction
+	{
+		Neutral,
+		Left,
+		Right
+	}
+}
